+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.3.0
 
+- Added the ability to show current tenant information.
 - Fixed a bug where the application icon wasn't reset after opening the manifest.
 
 ## 0.2.0